IAT-2241: Updated the application configuration to work with the new config model in common.
