+++ build.gradle
@@ -63,7 +63,7 @@ configurations {
 }
 
 dependencies {
-    compile 'org.opentestsystem.ap:ap-common:0.4.34'
+    compile 'org.opentestsystem.ap:ap-common:0.4.35-SNAPSHOT'
 
     compile 'org.springframework.boot:spring-boot-starter-actuator'
     compile 'org.springframework.boot:spring-boot-starter-security'
