+++ src/main/java/org/opentestsystem/ap/ivs/IVSApplication.java
@@ -15,21 +15,9 @@ package org.opentestsystem.ap.ivs;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.FilterType;
-import springfox.documentation.swagger2.annotations.EnableSwagger2;
 
-@EnableSwagger2
 @SpringBootApplication
-@ComponentScan(
-    basePackages = {"org.opentestsystem.ap.ivs", "org.opentestsystem.ap.common"},
-    excludeFilters = {
-        @ComponentScan.Filter(type = FilterType.REGEX, pattern = "org\\.opentestsystem\\.ap\\.common\\.datastore\\..*"),
-        @ComponentScan.Filter(type = FilterType.REGEX, pattern = "org\\.opentestsystem\\.ap\\.common\\" +
-            ".itembank\\.ItemManagerImpl"),
-        @ComponentScan.Filter(type = FilterType.REGEX, pattern = "org\\.opentestsystem\\.ap\\.common\\" +
-            ".bulkupdate\\..*")
-    }
-)
+@ComponentScan(basePackages = {"org.opentestsystem.ap.ivs", "org.opentestsystem.ap.common"})
 public class IVSApplication {
 
     static {
