--- src/test/java/org/opentestsystem/ap/ivs/config/GitlabClientStub.java
@@ -1,40 +1,0 @@
-/*
- * Copyright 2017 Regents of the University of California. Licensed under the Educational Community License, Version
- * 2.0 (the "license"); you may not use this file except in compliance with the License. You may obtain a copy of the
- * license at
- *
- * https://opensource.org/licenses/ECL-2.0
- *
- * Unless required under applicable law or agreed to in writing, software distributed under the License is distributed
- * in an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
- * specific language governing permissions and limitations under the license.
- */
-
-package org.opentestsystem.ap.ivs.config;
-
-import javax.annotation.PostConstruct;
-
-import org.gitlab4j.api.GitLabApi;
-import org.gitlab4j.api.models.Namespace;
-import org.opentestsystem.ap.common.client.GitlabClient;
-import org.opentestsystem.ap.common.config.ItemBankProperties;
-
-import static org.mockito.Mockito.mock;
-
-/**
- * Replaces the {@link GitlabClient} when running with the
- * itembank.test.enabled property is set to true.
- */
-public class GitlabClientStub extends GitlabClient {
-
-    public GitlabClientStub(ItemBankProperties gitlabProperties) {
-        super(gitlabProperties);
-    }
-
-    @Override
-    @PostConstruct
-    public void initialize() {
-        setGitLabApi(mock(GitLabApi.class));
-        setGroup(mock(Namespace.class));
-    }
-}
