--- src/test/java/org/opentestsystem/ap/ivs/config/NoItemBankConfig.java
@@ -1,58 +1,0 @@
-/*
- * Copyright 2017 Regents of the University of California. Licensed under the Educational Community License, Version
- * 2.0 (the "license"); you may not use this file except in compliance with the License. You may obtain a copy of the
- * license at
- *
- * https://opensource.org/licenses/ECL-2.0
- *
- * Unless required under applicable law or agreed to in writing, software distributed under the License is distributed
- * in an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
- * specific language governing permissions and limitations under the license.
- */
-package org.opentestsystem.ap.ivs.config;
-
-import org.mockito.Mockito;
-import org.opentestsystem.ap.common.client.GitlabClient;
-import org.opentestsystem.ap.common.config.ItemBankProperties;
-import org.opentestsystem.ap.common.datastore.DataStoreAttachmentManager;
-import org.opentestsystem.ap.common.datastore.DataStoreDataManager;
-import org.opentestsystem.ap.common.datastore.DataStoreUtility;
-import org.opentestsystem.ap.common.itembank.IdGenerator;
-import org.opentestsystem.ap.common.model.ItemFactory;
-import org.opentestsystem.ap.common.util.RandomIdGenerator;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-/**
- * No ItemBank configuration means we are running under a context where we do not have what is need to connect to an
- * actual Gitlab instance. This configuration replaces the runtime configuration
- */
-@Configuration
-@ConditionalOnProperty(value = "itembank.test.enabled", havingValue = "true", matchIfMissing = false)
-public class NoItemBankConfig {
-
-    private final ItemBankProperties props;
-
-    @Autowired
-    public NoItemBankConfig(ItemBankProperties props) {
-        this.props = props;
-    }
-
-    @Bean
-    public GitlabClient gitlabClient() {
-        return new GitlabClientStub(props);
-    }
-
-    @Bean
-    public IdGenerator itemIdGenerator() {
-        return new RandomIdGenerator(props.getIdMinValue(), props.getIdMaxValue());
-    }
-
-    @Bean
-    public ItemFactory itemFactory() {
-        return new ItemFactory();
-    }
-
-}
