--- src/test/java/org/opentestsystem/ap/ivs/config/NoRedisSessionConfig.java
@@ -1,62 +1,0 @@
-/*
- * Copyright 2017 Regents of the University of California. Licensed under the Educational Community License, Version
- * 2.0 (the "license"); you may not use this file except in compliance with the License. You may obtain a copy of the
- * license at
- *
- * https://opensource.org/licenses/ECL-2.0
- *
- * Unless required under applicable law or agreed to in writing, software distributed under the License is distributed
- * in an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for
- * specific language governing permissions and limitations under the license.
- */
-package org.opentestsystem.ap.ivs.config;
-
-import org.opentestsystem.ap.common.config.ItemBankProperties;
-import org.opentestsystem.ap.common.itembank.IdGenerator;
-import org.opentestsystem.ap.common.model.ItemFactory;
-import org.opentestsystem.ap.common.util.RandomIdGenerator;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
-import org.springframework.cache.CacheManager;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.cache.RedisCacheManager;
-import org.springframework.data.redis.core.RedisTemplate;
-
-import static org.mockito.Mockito.mock;
-
-/**
- * Disabled by default. If the gitlab.enabled property is set to false only then is
- * this configuration loaded.
- */
-@Configuration
-@ConditionalOnProperty(value = "itembank.test.enabled", havingValue = "true", matchIfMissing = false)
-public class NoRedisSessionConfig {
-
-    public final ItemBankProperties props;
-
-    @Autowired
-    public NoRedisSessionConfig(ItemBankProperties props) {
-        this.props = props;
-    }
-
-    @Bean
-    public IdGenerator itemIdGenerator() {
-        return new RandomIdGenerator(props.getIdMinValue(), props.getIdMaxValue());
-    }
-
-    @Bean
-    public ItemFactory itemFactory() {
-        return new ItemFactory();
-    }
-
-    @Bean
-    public RedisTemplate redisTemplate() {
-        return mock(RedisTemplate.class);
-    }
-
-    @Bean
-    public CacheManager cacheManager(RedisTemplate redisTemplate) {
-        return mock(RedisCacheManager.class);
-    }
-}
